extracted one-off generator into separate class
